requirements.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
fastapi==0.70.0
Flask==2.0.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.2
pydantic==1.8.2
python-dotenv==0.19.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
Werkzeug==2.0.1
